+++ opa-datafilter-jpa-spring-boot-starter/src/main/java/com/github/jferrater/opadatafilterjpaspringbootstarter/query/TypedQueryBuilder.java
@@ -99,8 +99,14 @@ public class TypedQueryBuilder<T> {
             case "<":
                 predicate = criteriaBuilder.lessThan(root.get(attributeName), Long.valueOf(rightExpression));
                 break;
+            case "<=":
+                predicate = criteriaBuilder.lessThanOrEqualTo(root.get(attributeName), Long.valueOf(rightExpression));
+                break;
             case ">":
                 predicate = criteriaBuilder.greaterThan(root.get(attributeName), Long.valueOf(rightExpression));
+                break;
+            case ">=":
+                predicate = criteriaBuilder.greaterThanOrEqualTo(root.get(attributeName), Long.valueOf(rightExpression));
                 break;
             default:
                 LOGGER.warn("Criteria equivalent of SQL operator '{}' is not yet supported", operator);