+++ opa-datafilter-mongo-spring-boot-starter/src/main/java/com/github/jferrater/opadatafiltermongospringbootstarter/query/AstToMongoDBQuery.java
@@ -100,10 +100,16 @@ public class AstToMongoDBQuery {
                 criteria = criteria.is(rightExpression);
                 break;
             case "<":
-                criteria = criteria.lt(Long.valueOf(rightExpression));
+                criteria = criteria.lt(Integer.valueOf(rightExpression));
+                break;
+            case "<=":
+                criteria = criteria.lte(Integer.valueOf(rightExpression));
                 break;
             case ">":
-                criteria = criteria.gt(Long.valueOf(rightExpression));
+                criteria = criteria.gt(Integer.valueOf(rightExpression));
+                break;
+            case ">=":
+                criteria = criteria.gte(Integer.valueOf(rightExpression));
                 break;
             default:
                 LOGGER.warn("Mongo DB criteria equivalent of SQL operator '{}' is not yet supported", operator);